# docs/swagger.yaml
basePath: /
definitions:
...
  models.AddPostRequest:
    properties:
      link:
        type: string
    required:
    - link
    type: object
  models.AddPostResponse:
    properties:
      content_id:
        type: string
      media_count:
        type: integer
      message:
        type: string
      processing_status:
        $ref: '#/definitions/models.PostStatus'
      status:
        type: string
    type: object
...
  models.PostListItem:
    properties:
      added_at:
        type: string
      content_id:
        type: string
      link:
        type: string
      media_count:
        type: integer
      status:
        $ref: '#/definitions/models.PostStatus'
    type: object
  models.PostListResponse:
    properties:
      limit:
        type: integer
      links:
        items:
          $ref: '#/definitions/models.PostListItem'
        type: array
      page:
        type: integer
      total:
        type: integer
    type: object
  models.PostStatus:
    enum:
    - pending
    - processing
    - completed
    - failed
    type: string
    x-enum-varnames:
    - PostStatusPending
    - PostStatusProcessing
    - PostStatusCompleted
    - PostStatusFailed
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A Go-based microservice for planning and scheduling YouTube live streams.
    Helps content creators organize their streaming schedule and manage content.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: St. Planer - YouTube Stream Planner API
  version: "1.0"
paths:
...
  /api/v1/media/grab:
    post:
      consumes:
      - application/json
      description: Add a new Telegram post link or YouTube video URL to download media.
        Automatically detects the platform and processes accordingly.
      parameters:
      - description: Post link (Telegram or YouTube)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.AddPostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AddPostResponse'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.AddPostResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add a new Telegram or YouTube link for processing
      tags:
      - media
...
  /api/v1/media/list:
    get:
      consumes:
      - application/json
      description: Retrieve list of all previously processed Telegram and YouTube
        links
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Sort order
        enum:
        - created_at_desc
        - created_at_asc
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostListResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get list of processed posts
      tags:
      - media
...
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'JWT Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"